Render lines at the right place (partial work)
